+++ build_tools/aws-sdk-code-generator/lib/aws-sdk-code-generator/views/client_api_module.rb
@@ -192,6 +192,14 @@ module AwsSdkCodeGenerator
                 a << { key: k.inspect, value: v.inspect }
                 a
               end
+            # endpoint trait cannot be co-exist with endpoint discovery
+            elsif operation.key?('endpoint')
+              # endpoint trait per operation, cannot be enabled with endpoint discovery
+              o.endpoint_trait = true
+              o.endpoint_pattern = operation['endpoint'].inject([]) do |a, (k, v)|
+                a << { key: k.inspect, value: v.inspect }
+                a
+              end
             end
             o.authorizer = operation['authorizer'] if operation.key?('authorizer')
             o.authtype = operation['authtype'] if operation.key?('authtype')
@@ -290,7 +298,7 @@ module AwsSdkCodeGenerator
         line += shape_ref_eventheader(ref)
         line += shape_ref_location(ref)
         line += shape_ref_location_name(member_name, ref)
-        line += shape_ref_metadata(ref)
+        line += shape_ref_metadata(ref, member_name)
         line += ")"
         line
       end
@@ -359,9 +367,12 @@ module AwsSdkCodeGenerator
         location_name ? ", location_name: #{location_name.inspect}" : ""
       end
 
-      def shape_ref_metadata(member_ref)
+      def shape_ref_metadata(member_ref, member_name)
         metadata = member_ref.inject({}) do |hash, (key, value)|
           hash[key] = value unless SKIP_TRAITS.include?(key)
+          if key == 'hostLabel'
+            hash['hostLabelName'] = member_name
+          end
           hash
         end
         if metadata.empty?
@@ -480,6 +491,12 @@ module AwsSdkCodeGenerator
 
         # @return [String,nil]
         attr_accessor :authtype
+
+        # @return [Boolean]
+        attr_accessor :endpoint_trait
+
+        # @return [Array]
+        attr_accessor :endpoint_pattern
 
         # APIG only
         # @return [Boolean]